# packages/types/openapi/cetem-qc-v1.yaml
openapi: 3.1.0
info:
  title: CETEM-QC API
  version: 1.0.0
  description: Versioned public HTTP contract. Internal identifiers and error codes are English.
servers:
  - url: /api/v1
paths:
...
  /tasks/{taskId}/report-candidates:
    post:
      operationId: generateReportCandidate
      summary: Generate a Word report candidate from the current audit workflow inputs
      description: >-
        Responsable only. The body carries a client-generated attemptId only; actor and dates come from the session and
        the server clock. Generates one Word document, stores it privately and returns the candidate, which is bound
        immutably to the audit revision, the confirmed summary and the conformity decision. A candidate is never
        official in this story. The same attemptId for the same task replays the stored outcome (200 for a ready
        candidate, 502 for a failed one) without generating again. A malformed, unknown or another team's task, and an
        own-team task without accepted submission, all get the same 404.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCandidateRequest'
      responses:
        '200':
          description: Replay of an attempt that already produced a ready candidate (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportCandidate'
        '201':
          description: Generated candidate (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportCandidate'
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '403': { description: Responsable role required (FORBIDDEN), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: Malformed, unknown or another team's task, or no accepted submission (TASK_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '409': { description: Summary not confirmed (SUMMARY_NOT_CONFIRMED), no current conformity decision (CONFORMITY_NOT_DECIDED), attempt used by another task (REPORT_ATTEMPT_CONFLICT) inputs changed during generation (REPORT_INPUTS_CHANGED) or an official report already designated (REPORT_OFFICIAL_DESIGNATED), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '422': { description: Missing or invalid attemptId, or extra property (VALIDATION_FAILED), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: Evidence inconsistent or candidate not recorded (INTERNAL_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '502': { description: Generation or storage failed; the attempt can be retried with a new attemptId (REPORT_GENERATION_FAILED), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    get:
      operationId: listReportCandidates
      summary: List the Word and PDF report candidates of a task, newest first
      description: >-
        Responsable only. The status of each candidate is derived on every read: a ready candidate whose bound summary
        or conformity decision is no longer current reads as outdated. A malformed, unknown or another team's task, and
        an own-team task without accepted submission, all get the same 404.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Candidates, newest first (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportCandidateList'
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '403': { description: Responsable role required (FORBIDDEN), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: Malformed, unknown or another team's task, or no accepted submission (TASK_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: Candidates could not be loaded (INTERNAL_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /tasks/{taskId}/pdf-files/{fileId}/report-candidate:
    post:
      operationId: createReportCandidateFromPdf
      summary: Create a report candidate from a ready manual PDF
      description: >-
        Responsable only. The body carries a client-generated attemptId only. Attaches a ready (validated, scanned) stored
        PDF of an eligible task as a report candidate bound immutably to the audit revision, the confirmed summary and the
        conformity decision. The candidate is never official and the PDF content is not inspected. The same attemptId for
        the same task and file replays the stored candidate (200). A malformed, unknown or another team's task or file all
        get the same 404.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportFromPdfRequest'
      responses:
        '200':
          description: Replay of an attempt that already attached this file (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportCandidate'
        '201':
          description: Attached candidate (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportCandidate'
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '403': { description: Responsable role required (FORBIDDEN), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: Malformed, unknown or another team's task or file (TASK_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '409': { description: Summary not confirmed (SUMMARY_NOT_CONFIRMED), no current conformity decision (CONFORMITY_NOT_DECIDED), file not ready (REPORT_FILE_NOT_READY), file already a candidate for the current data (REPORT_FILE_ALREADY_ATTACHED) attempt used elsewhere (REPORT_ATTEMPT_CONFLICT) or an official report already designated (REPORT_OFFICIAL_DESIGNATED), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '422': { description: Missing or invalid attemptId, or extra property (VALIDATION_FAILED), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: Evidence inconsistent or candidate not recorded (INTERNAL_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /tasks/{taskId}/report-candidates/{candidateId}/designate:
    post:
      operationId: designateReportCandidate
      summary: Designate one ready report candidate as the official report of the control
      description: >-
        Responsable only, owning team only. The body is an empty object. Designates one ready, current candidate (Word
        or PDF) as the single official report of the task; the choice is final and nothing can replace it. The server
        rechecks in one transaction that the candidate is ready and still bound to the current confirmed summary,
        conformity decision and accepted submission, and that a PDF file is still ready. Designating the same candidate
        again returns the stored official report (200). Either conformity outcome can be designated. A malformed,
        unknown or another team's task or candidate all get the same 404.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: candidateId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportDesignateRequest'
      responses:
        '200':
          description: Replay of a designation of the same candidate (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficialReport'
        '201':
          description: Designated official report (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficialReport'
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '403': { description: Responsable role required (FORBIDDEN), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: Malformed, unknown or another team's task or candidate (TASK_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '409': { description: Candidate not ready (REPORT_CANDIDATE_NOT_READY), candidate outdated (REPORT_CANDIDATE_OUTDATED) or another candidate already official (REPORT_ALREADY_OFFICIAL), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '422': { description: Body is not an empty object (VALIDATION_FAILED), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: Evidence inconsistent or designation not recorded (INTERNAL_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /tasks/{taskId}/official-report:
    get:
      operationId: getOfficialReport
      summary: Read the official report of a task
      description: >-
        Responsable only, owning team only. A task without an official report, a malformed, unknown or another team's
        task, and an own-team task without accepted submission all get the same 404.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The official report (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficialReport'
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '403': { description: Responsable role required (FORBIDDEN), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: No official report, or malformed, unknown or another team's task (TASK_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: Official report could not be loaded (INTERNAL_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /tasks/{taskId}/report-candidates/{candidateId}/file:
    get:
      operationId: downloadReportCandidateFile
      summary: Download the stored Word or PDF file of a candidate for inspection
      description: >-
        Responsable only, owning team only. Available for a candidate whose outcome retained a file (ready or outdated).
        The response is an attachment with Cache-Control no-store and X-Content-Type-Options nosniff. Any other case,
        including a failed or generating candidate and another team's candidate, gets the same 404 as an unknown task.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: candidateId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The Word document or the uploaded PDF (Content-Disposition attachment, Cache-Control no-store, X-Content-Type-Options nosniff)
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema: { type: string, format: binary }
            application/pdf:
              schema: { type: string, format: binary }
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '403': { description: Responsable role required (FORBIDDEN), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: Malformed, unknown, other team's, failed or generating candidate, or unknown task (TASK_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: The stored file could not be read (INTERNAL_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    ReportCandidateRequest:
      type: object
      additionalProperties: false
      required: [attemptId]
      properties:
        attemptId: { type: string, format: uuid }
    ReportFromPdfRequest:
      type: object
      additionalProperties: false
      required: [attemptId]
      properties:
        attemptId: { type: string, format: uuid }
    ReportDesignateRequest:
      type: object
      additionalProperties: false
      properties: {}
    ReportCandidateStatus:
      type: string
      enum: [generating, ready, failed, outdated, official, superseded]
...
    ReportCandidate:
      type: object
      description: >-
        A report candidate, generated Word (template set, source null) or uploaded PDF (template null, source set). It is never official in this story, and no official designation property exists.
      additionalProperties: false
      required: [id, attemptId, origin, status, requestedAt, requestedBy, bindings, template, source, file, failureClass]
      properties:
        id: { type: string, format: uuid }
        attemptId: { type: string, format: uuid }
        origin:
          type: string
          enum: [generated-word, uploaded-pdf]
        status:
          $ref: '#/components/schemas/ReportCandidateStatus'
        requestedAt: { type: string, format: date-time }
        requestedBy:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
        bindings:
          type: object
          additionalProperties: false
          required: [auditRevision, summaryId, summaryVersion, conformityDecisionId, conformityOutcome]
          properties:
            auditRevision: { type: integer, minimum: 1 }
            summaryId: { type: string, format: uuid }
            summaryVersion: { type: integer, minimum: 1 }
            conformityDecisionId: { type: string, format: uuid }
            conformityOutcome:
              $ref: '#/components/schemas/ConformityOutcome'
        template:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [id, version]
              properties:
                id: { type: string }
                version: { type: string }
        source:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [fileId, scanResult]
              properties:
                fileId: { type: string, format: uuid }
                scanResult:
                  type: string
                  enum: [clean, not-performed]
        file:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [name, byteSize, sha256]
              properties:
                name: { type: string }
                byteSize: { type: integer, minimum: 0 }
                sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        failureClass:
          oneOf:
            - type: 'null'
            - type: string
              enum: [generation-failed, storage-failed]
    ReportCandidateList:
      type: object
      additionalProperties: false
      required: [candidates]
      properties:
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/ReportCandidate'
...
    ApiError:
      type: object
      description: >-
        Error envelope. Codes include VALIDATION_ERROR, AUTHENTICATION_FAILED, FORBIDDEN, TASK_NOT_FOUND,
        PAYLOAD_TOO_LARGE, IDEMPOTENCY_KEY_REUSED, AUDIT_NOT_ACCEPTED, REPLACEMENT_ALREADY_EXISTS,
        TASK_ASSIGNEE_UNAVAILABLE, AI_UNAVAILABLE, SUMMARY_ALREADY_CONFIRMED, SUMMARY_CONFIRMED, SUMMARY_NOT_CONFIRMED, SUMMARY_DESIGNATED, CONFORMITY_ALREADY_DECIDED, CONFORMITY_NOT_DECIDED, REPORT_GENERATION_FAILED,
        REPORT_INPUTS_CHANGED, REPORT_ATTEMPT_CONFLICT, REPORT_FILE_NOT_READY, REPORT_FILE_ALREADY_ATTACHED, REPORT_CANDIDATE_NOT_READY, REPORT_CANDIDATE_OUTDATED, REPORT_ALREADY_OFFICIAL, REPORT_OFFICIAL_DESIGNATED, FILE_TOO_LARGE, UNSUPPORTED_FILE_TYPE, FILE_STORAGE_FAILED,
        FILE_ATTEMPT_CONFLICT, FILE_NOT_RESCANNABLE and INTERNAL_ERROR.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                required: [path, message]
                properties:
                  path:
                    type: string
                  message:
                    type: string